refactor: map-topology SoT and densify map/Flame/session tests (#4642)
Reuse shortcut-host OW+NW fixtures in mapViewDataProvider tests and
concern-split Slice B suites so each file stays under 330 lines.
