+++ app/test/map_view_provider_test.dart
@@ -1,4 +1,3 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter_test/flutter_test.dart';
 
@@ -15,6 +14,8 @@ import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hive/hive.dart';
 
+import 'province_shortcut_host_golden_game_service.dart';
+
 class _FakeGameService extends GameService {
   _FakeGameService(super.box, super.adapter);
 
@@ -24,109 +25,14 @@ class _FakeGameService extends GameService {
   }
 }
 
-/// Minimal OW/NW tile maps + topology for [mapViewDataProvider] integration tests.
-class _GameServiceWithMinimalMap extends GameService {
-  _GameServiceWithMinimalMap(super.box, super.adapter);
-
-  static final Map<String, MapTopology> _topologyByRegion = {
-    'oldWorld': MapTopology(
-      nodes: const [
-        TopologyNode(
-          id: 'p1',
-          regionId: 'oldWorld',
-          type: TopologyNodeType.province,
-        ),
-        TopologyNode(
-          id: 's1',
-          regionId: 'oldWorld',
-          type: TopologyNodeType.seaZone,
-        ),
-      ],
-      edges: const [TopologyEdge(id1: 'p1', id2: 's1')],
-    ),
-    'newWorld': MapTopology(
-      nodes: const [
-        TopologyNode(
-          id: 'p1',
-          regionId: 'newWorld',
-          type: TopologyNodeType.province,
-        ),
-        TopologyNode(
-          id: 's1',
-          regionId: 'newWorld',
-          type: TopologyNodeType.seaZone,
-        ),
-      ],
-      edges: const [TopologyEdge(id1: 'p1', id2: 's1')],
-    ),
-  };
-
-  static final Map<String, TileMapResult> _tileMapByRegion = {
-    'oldWorld': TileMapResult(
-      width: 2,
-      height: 2,
-      grid: [
-        ['p1', 's1'],
-        ['s1', 's1'],
-      ],
-    ),
-    'newWorld': TileMapResult(
-      width: 2,
-      height: 2,
-      grid: [
-        ['p1', 's1'],
-        ['s1', 's1'],
-      ],
-    ),
-  };
-
-  static final MapTopology _combinedTopology = MapTopology(
-    nodes: const [
-      TopologyNode(
-        id: 'oldWorld|p1',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.province,
-      ),
-      TopologyNode(
-        id: 'oldWorld|s1',
-        regionId: 'oldWorld',
-        type: TopologyNodeType.seaZone,
-      ),
-      TopologyNode(
-        id: 'newWorld|p1',
-        regionId: 'newWorld',
-        type: TopologyNodeType.province,
-      ),
-      TopologyNode(
-        id: 'newWorld|s1',
-        regionId: 'newWorld',
-        type: TopologyNodeType.seaZone,
-      ),
-    ],
-    edges: const [
-      TopologyEdge(id1: 'oldWorld|p1', id2: 'oldWorld|s1'),
-      TopologyEdge(id1: 'newWorld|p1', id2: 'newWorld|s1'),
-    ],
+GameService _shortcutHostMapService(Box<dynamic> gamesBox) {
+  return ProvinceShortcutHostGoldenGameService(
+    gamesBox,
+    GameSaveAdapter(),
+    gameId: 'g_map',
+    includeNewWorld: true,
+    useCoastalTileMap: true,
   );
-
-  @override
-  ({
-    MapTopology combinedTopology,
-    Map<String, TileMapResult> tileMapByRegion,
-    Map<String, MapTopology> topologyByRegion,
-    List<WarpLink>? warpLinks,
-  })?
-  getMapData(String gameId) {
-    if (gameId != 'g_map') {
-      return null;
-    }
-    return (
-      combinedTopology: _combinedTopology,
-      tileMapByRegion: _tileMapByRegion,
-      topologyByRegion: _topologyByRegion,
-      warpLinks: null,
-    );
-  }
 }
 
 void main() {
@@ -227,7 +133,7 @@ void main() {
           currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
           gamesBoxProvider.overrideWith((ref) => gamesBox),
           gameServiceProvider.overrideWith(
-            (ref) => _GameServiceWithMinimalMap(gamesBox, GameSaveAdapter()),
+            (ref) => _shortcutHostMapService(gamesBox),
           ),
         ],
       );
@@ -240,74 +146,71 @@ void main() {
     },
   );
 
-  test(
-    'mapViewDataProvider reveals all tiles in global observe mode',
-    () {
-      const unrevealedTile = 'oldWorld|p1|0|0';
-      final game = Game(
-        id: 'g_map',
-        worldState: WorldState(
-          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-          oldWorld: RegionData(
-            provinces: const [
-              Province(
-                id: 'oldWorld|p1',
-                regionId: 'oldWorld',
-                displayName: 'OW P1',
-                ownerId: 'gp1',
-              ),
-            ],
-            units: const [],
-          ),
-          newWorld: const RegionData(),
-          tileKeysByRegionAndProvince: const {
-            'oldWorld': {
-              'oldWorld|p1': [unrevealedTile],
-            },
-          },
-          playerVisibilityByTile: const {
-            'gp1': {unrevealedTile: 'unknown'},
-          },
+  test('mapViewDataProvider reveals all tiles in global observe mode', () {
+    const unrevealedTile = 'oldWorld|p1|0|0';
+    final game = Game(
+      id: 'g_map',
+      worldState: WorldState(
+        turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
+        oldWorld: RegionData(
+          provinces: const [
+            Province(
+              id: 'oldWorld|p1',
+              regionId: 'oldWorld',
+              displayName: 'OW P1',
+              ownerId: 'gp1',
+            ),
+          ],
+          units: const [],
         ),
-        players: const [
-          Player(id: 'gp1', displayName: 'Human', isHuman: true, treasury: 0),
-          Player(id: 'gp2', displayName: 'France', isHuman: false),
-        ],
-      );
+        newWorld: const RegionData(),
+        tileKeysByRegionAndProvince: const {
+          'oldWorld': {
+            'oldWorld|p1': [unrevealedTile],
+          },
+        },
+        playerVisibilityByTile: const {
+          'gp1': {unrevealedTile: 'unknown'},
+        },
+      ),
+      players: const [
+        Player(id: 'gp1', displayName: 'Human', isHuman: true, treasury: 0),
+        Player(id: 'gp2', displayName: 'France', isHuman: false),
+      ],
+    );
 
-      final container = ProviderContainer(
-        overrides: [
-          currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
-          gamesBoxProvider.overrideWith((ref) => gamesBox),
-          gameServiceProvider.overrideWith(
-            (ref) => _GameServiceWithMinimalMap(gamesBox, GameSaveAdapter()),
-          ),
-        ],
-      );
-      addTearDown(container.dispose);
+    final container = ProviderContainer(
+      overrides: [
+        currentGameProvider.overrideWith(() => CurrentGameNotifier(game)),
+        gamesBoxProvider.overrideWith((ref) => gamesBox),
+        gameServiceProvider.overrideWith(
+          (ref) => _shortcutHostMapService(gamesBox),
+        ),
+      ],
+    );
+    addTearDown(container.dispose);
 
-      final constrained = container.read(mapViewDataProvider);
-      expect(constrained, isNotNull);
-      expect(
-        constrained!.oldWorld.cellAt(0, 0).visibility,
-        TileVisibility.unrevealed,
-      );
+    final constrained = container.read(mapViewDataProvider);
+    expect(constrained, isNotNull);
+    expect(
+      constrained!.oldWorld.cellAt(0, 0).visibility,
+      TileVisibility.unrevealed,
+    );
 
-      final observe = container.read(observeSessionProvider.notifier);
-      final handedOff = observe.applyObserveHandoffIfNeeded(game);
-      container.read(currentGameProvider.notifier).setGame(handedOff);
-      observe.setModeGlobal();
+    final observe = container.read(observeSessionProvider.notifier);
+    final handedOff = observe.applyObserveHandoffIfNeeded(game);
+    container.read(currentGameProvider.notifier).setGame(handedOff);
+    observe.setModeGlobal();
 
-      final globalObserve = container.read(mapViewDataProvider);
-      expect(globalObserve, isNotNull);
-      for (final cell in globalObserve!.oldWorld.cells) {
-        expect(cell.visibility, TileVisibility.visible);
-      }
-      for (final cell in globalObserve.newWorld.cells) {
-        expect(cell.visibility, TileVisibility.visible);
-      }
-    },
-  );
+    final globalObserve = container.read(mapViewDataProvider);
+    expect(globalObserve, isNotNull);
+    for (final cell in globalObserve!.oldWorld.cells) {
+      expect(cell.visibility, TileVisibility.visible);
+    }
+    for (final cell in globalObserve.newWorld.cells) {
+      expect(cell.visibility, TileVisibility.visible);
+    }
+  });
 
   test(
     'mapViewDataProvider includes non-human civilian markers in global observe',
@@ -318,16 +221,8 @@ void main() {
           turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
           oldWorld: RegionData(
             provinces: const [
-              Province(
-                id: 'oldWorld|p1',
-                regionId: 'oldWorld',
-                ownerId: 'gp1',
-              ),
-              Province(
-                id: 'oldWorld|p2',
-                regionId: 'oldWorld',
-                ownerId: 'gp2',
-              ),
+              Province(id: 'oldWorld|p1', regionId: 'oldWorld', ownerId: 'gp1'),
+              Province(id: 'oldWorld|p2', regionId: 'oldWorld', ownerId: 'gp2'),
             ],
             units: [
               Unit(
@@ -353,7 +248,7 @@ void main() {
           currentGameProvider.overrideWith(CurrentGameNotifier.new),
           gamesBoxProvider.overrideWith((ref) => gamesBox),
           gameServiceProvider.overrideWith(
-            (ref) => _GameServiceWithMinimalMap(gamesBox, GameSaveAdapter()),
+            (ref) => _shortcutHostMapService(gamesBox),
           ),
         ],
       );
@@ -370,8 +265,10 @@ void main() {
       container.read(currentGameProvider.notifier).setGame(handedOff);
       observe.setModeGlobal();
 
-      final markers =
-          container.read(mapViewDataProvider)!.oldWorld.civilianTileMarkers;
+      final markers = container
+          .read(mapViewDataProvider)!
+          .oldWorld
+          .civilianTileMarkers;
       expect(markers, hasLength(1));
       expect(markers.single.tileKey, 'oldWorld|p2|1|0');
     },
